Add a DogStatsd sink
This adds a very basic DogStatsdSink implementing the MetricSink
interface.
It has two main features:
1. Using the client at https://github.com/DataDog/datadog-go/statsd it
allows global tags to be assigned to all metrics sent into the sink
2. When necessary, it splices out the hostname from incoming metric
names, and optinally delegates host tagging to the Dogstatsd server. This allows
for much saner grouping by host in the Datadog backend.