Place the hostname in front of the type prefix
This makes it easier to organize metrics in a hierarchy. All metrics for
a specific host are under the hostname so instead of:

    <service>.<type>.<hostname> = consul.gauge.consul-box

You'll end up with metrics that will organize all metrics for a box
under a single identifier.

    <service>.<hostname>.<type> = consul.consul-box.gauge

This makes it easier to use things like `*` in tools like Grafana for
viewing metrics from Graphite.