Introduce unified URL-based sink factory function
The `NewMetricSinkFromURL` function allows a single string input to
configure any of the supported sink backends. The various pieces of the
URL are interpreted on a per-sink basis, and the generic `MetricSink`
interface is returned.

This is convenient for downstream users of go-metrics that want to
expose sink-agnostic configuration options, but for whom a more
complicated (e.g. HCL-based) config language is too heavy.

Currently the statsd, statsite, and inmem sinks are supported. Support
for the remaining sinks can be added as a follow up.