Expose the global default metrics
This should make it easier to test code that uses go-metrics. In tests the code can be
passed a different instance of Metrics, and in production code paths it can use the global
Metrics returned from Default().

Without this change patching the global can cause flaky tests when other tests attempt to
emit metrics at the same time.